Replace use of vg_datetime_to_millis to str_to_timestamptz
Remove DateMode as it's not longer needed
