+++ vegafusion-rt-datafusion/src/expression/compiler/builtin_functions/date_time/mod.rs
@@ -13,18 +13,15 @@ Functions for working with date-time values.
 See: https://vega.github.io/vega/docs/expressions/#datetime-functions
 */
 pub mod date_format;
-pub mod date_parsing;
 pub mod date_parts;
-pub mod datetime;
-pub mod time;
 pub mod date_to_timestamptz;
-pub mod timestamp_to_timestamptz;
+pub mod datetime;
 pub mod epoch_to_timestamptz;
 pub mod str_to_timestamptz;
+pub mod time;
+pub mod timestamp_to_timestamptz;
 
-use crate::expression::compiler::builtin_functions::date_time::date_parsing::{
-    datetime_strs_to_timestamp_millis, DateParseMode,
-};
+use crate::expression::compiler::builtin_functions::date_time::str_to_timestamptz::datetime_strs_to_timestamp_millis;
 use datafusion::arrow::array::{ArrayRef, Date32Array, Int64Array, StringArray};
 use datafusion::arrow::compute::{cast, unary};
 use datafusion::arrow::datatypes::{DataType, TimeUnit};
@@ -35,11 +32,7 @@ pub fn process_input_datetime(arg: &ArrayRef, default_input_tz: &chrono_tz::Tz) 
         DataType::Utf8 => {
             let array = arg.as_any().downcast_ref::<StringArray>().unwrap();
             cast(
-                &datetime_strs_to_timestamp_millis(
-                    array,
-                    DateParseMode::JavaScript,
-                    &Some(*default_input_tz),
-                ),
+                &datetime_strs_to_timestamp_millis(array, &Some(*default_input_tz)),
                 &DataType::Int64,
             )
             .expect("Failed to case timestamp to Int64")
